requirements.txt
# Core LangChain and LangGraph - 使用更新的兼容版本
langchain==0.1.20
langgraph==0.0.65
langchain-openai==0.1.8
langchain-community==0.0.38

# LLM and API clients
openai==1.35.0
anthropic==0.25.0

# Search and data retrieval - 修复版本冲突
tavily-python==0.3.3
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
selenium==4.15.0

# Document processing
python-pptx==0.6.23
python-docx==1.0.1
PyPDF2==3.0.1
pymupdf==1.23.26

# Data processing
pandas==2.1.4
numpy==1.24.3
pydantic==2.7.0

# Utilities
python-dotenv==1.0.0
typing-extensions==4.9.0
aiohttp==3.9.5

# 明确指定tiktoken版本以避免冲突
tiktoken==0.5.2

# Optional: for enhanced functionality
yfinance==0.2.28
feedparser==6.0.11